split CdcPartitionReaderTest
