+++ airbyte-cdk/bulk/toolkits/extract-cdc/src/test/kotlin/io/airbyte/cdk/read/cdc/CdcPartitionReaderTest.kt
@@ -4,18 +4,10 @@
 
 package io.airbyte.cdk.read.cdc
 
-import com.alibaba.dcm.DnsCacheManipulator
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
 import com.fasterxml.jackson.databind.node.ArrayNode
 import com.fasterxml.jackson.databind.node.ObjectNode
-import com.mongodb.client.MongoClient
-import com.mongodb.client.MongoClients
-import com.mongodb.client.MongoCollection
-import com.mongodb.client.MongoDatabase
-import com.mongodb.client.model.Aggregates
-import com.mongodb.client.model.Filters
-import com.mongodb.client.model.Updates
 import io.airbyte.cdk.ClockFactory
 import io.airbyte.cdk.StreamIdentifier
 import io.airbyte.cdk.command.OpaqueStateValue
@@ -31,39 +23,18 @@ import io.airbyte.cdk.read.PartitionReadCheckpoint
 import io.airbyte.cdk.read.PartitionReader
 import io.airbyte.cdk.read.Stream
 import io.airbyte.cdk.read.StreamRecordConsumer
-import io.airbyte.cdk.testcontainers.TestContainerFactory
 import io.airbyte.cdk.util.Jsons
 import io.airbyte.protocol.models.v0.AirbyteRecordMessage
 import io.airbyte.protocol.models.v0.StreamDescriptor
-import io.debezium.connector.mongodb.MongoDbConnector
-import io.debezium.connector.mongodb.ResumeTokens
-import io.debezium.connector.mysql.MySqlConnector
-import io.debezium.connector.postgresql.PostgresConnector
 import io.debezium.document.DocumentReader
 import io.debezium.document.DocumentWriter
 import io.debezium.relational.history.HistoryRecord
-import io.debezium.testing.testcontainers.MongoDbReplicaSet
-import java.sql.Connection
-import java.sql.Statement
 import java.time.Duration
-import java.time.Instant
-import java.util.regex.Pattern
-import kotlin.random.Random
-import kotlin.random.nextInt
 import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withTimeout
-import org.apache.kafka.connect.source.SourceRecord
-import org.bson.BsonDocument
-import org.bson.BsonTimestamp
-import org.bson.Document
-import org.bson.conversions.Bson
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
-import org.postgresql.replication.LogSequenceNumber
-import org.testcontainers.containers.MySQLContainer
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.utility.DockerImageName
 
 /**
  * This test class verifies that the [CdcPartitionReader] is able to correctly start and stop the
@@ -71,7 +42,7 @@ import org.testcontainers.utility.DockerImageName
  * integration test and this class is subclassed for multiple Debezium implementations which connect
  * to a corresponding testcontainer data source.
  */
-sealed class CdcPartitionReaderTest<T : Comparable<T>, C : AutoCloseable>(
+abstract class CdcPartitionReaderTest<T : Comparable<T>, C : AutoCloseable>(
     namespace: String?,
     val heartbeat: Duration = Duration.ofMillis(100),
     val timeout: Duration = Duration.ofSeconds(10),
@@ -313,401 +284,5 @@ sealed class CdcPartitionReaderTest<T : Comparable<T>, C : AutoCloseable>(
                 },
             )
         return DebeziumState(offset, schemaHistory)
-    }
-}
-
-class CdcPartitionReaderMySQLTest :
-    CdcPartitionReaderTest<CdcPartitionReaderMySQLTest.Position, MySQLContainer<*>>(
-        namespace = "test",
-    ) {
-
-    data class Position(val file: String, val pos: Long) : Comparable<Position> {
-        override fun compareTo(other: Position): Int =
-            file.compareTo(other.file).takeUnless { it == 0 } ?: pos.compareTo(other.pos)
-    }
-
-    override fun createContainer(): MySQLContainer<*> {
-        val dockerImageName: DockerImageName =
-            DockerImageName.parse(DOCKER_IMAGE_NAME).asCompatibleSubstituteFor(DOCKER_IMAGE_NAME)
-        val modifier: TestContainerFactory.ContainerModifier<MySQLContainer<*>> =
-            TestContainerFactory.newModifier("withRoot") { it.withUsername("root") }
-        return TestContainerFactory.exclusive(dockerImageName, modifier)
-    }
-
-    companion object {
-        const val DOCKER_IMAGE_NAME = "mysql:8.0"
-        init {
-            TestContainerFactory.register(DOCKER_IMAGE_NAME, ::MySQLContainer)
-        }
-    }
-
-    override fun MySQLContainer<*>.createStream() {
-        withStatement { it.execute("CREATE TABLE tbl (id INT AUTO_INCREMENT PRIMARY KEY, v INT)") }
-    }
-
-    override fun MySQLContainer<*>.insert12345() {
-        for (i in 1..5) {
-            withStatement { it.execute("INSERT INTO tbl (v) VALUES ($i)") }
-        }
-    }
-
-    override fun MySQLContainer<*>.update135() {
-        withStatement { it.execute("UPDATE tbl SET v = 6 WHERE id = 1") }
-        withStatement { it.execute("UPDATE tbl SET v = 7 WHERE id = 3") }
-        withStatement { it.execute("UPDATE tbl SET v = 8 WHERE id = 5") }
-    }
-
-    override fun MySQLContainer<*>.delete24() {
-        withStatement { it.execute("DELETE FROM tbl WHERE id = 2") }
-        withStatement { it.execute("DELETE FROM tbl WHERE id = 4") }
-    }
-
-    private fun <X> MySQLContainer<*>.withStatement(fn: (Statement) -> X): X =
-        createConnection("").use { connection: Connection ->
-            connection.createStatement().use { fn(it) }
-        }
-
-    override fun position(recordValue: DebeziumRecordValue): Position? {
-        val file: String =
-            recordValue.source["file"]?.takeIf { it.isTextual }?.asText() ?: return null
-        val pos: Long =
-            recordValue.source["pos"]?.takeIf { it.isIntegralNumber }?.asLong() ?: return null
-        return Position(file, pos)
-    }
-
-    override fun position(sourceRecord: SourceRecord): Position? {
-        val offset: Map<String, *> = sourceRecord.sourceOffset()
-        val file: String = offset["file"]?.toString() ?: return null
-        val pos: Long = offset["pos"] as? Long ?: return null
-        return Position(file, pos)
-    }
-
-    override fun MySQLContainer<*>.currentPosition(): Position =
-        withStatement { statement: Statement ->
-            statement.executeQuery("SHOW MASTER STATUS").use {
-                it.next()
-                Position(it.getString("File"), it.getLong("Position"))
-            }
-        }
-
-    override fun MySQLContainer<*>.syntheticInput(): DebeziumInput {
-        val position: Position = currentPosition()
-        val timestamp: Instant = Instant.now()
-        val key: ArrayNode =
-            Jsons.arrayNode().apply {
-                add(databaseName)
-                add(Jsons.objectNode().apply { put("server", databaseName) })
-            }
-        val value: ObjectNode =
-            Jsons.objectNode().apply {
-                put("ts_sec", timestamp.epochSecond)
-                put("file", position.file)
-                put("pos", position.pos)
-            }
-        val offset = DebeziumOffset(mapOf(key to value))
-        val state = DebeziumState(offset, schemaHistory = null)
-        val syntheticProperties: Map<String, String> =
-            DebeziumPropertiesBuilder()
-                .with(debeziumProperties())
-                .with("snapshot.mode", "recovery")
-                .withStreams(listOf())
-                .buildMap()
-        return DebeziumInput(syntheticProperties, state, isSynthetic = true)
-    }
-
-    override fun MySQLContainer<*>.debeziumProperties(): Map<String, String> =
-        DebeziumPropertiesBuilder()
-            .withDefault()
-            .withConnector(MySqlConnector::class.java)
-            .withDebeziumName(databaseName)
-            .withHeartbeats(heartbeat)
-            .with("include.schema.changes", "false")
-            .with("connect.keep.alive.interval.ms", "1000")
-            .withDatabase("hostname", host)
-            .withDatabase("port", firstMappedPort.toString())
-            .withDatabase("user", username)
-            .withDatabase("password", password)
-            .withDatabase("dbname", databaseName)
-            .withDatabase("server.id", Random.Default.nextInt(5400..6400).toString())
-            .withDatabase("include.list", databaseName)
-            .withOffset()
-            .withSchemaHistory()
-            .with("snapshot.mode", "when_needed")
-            .withStreams(listOf(stream))
-            .buildMap()
-}
-
-class CdcPartitionReaderPostgresTest :
-    CdcPartitionReaderTest<LogSequenceNumber, PostgreSQLContainer<*>>(
-        namespace = "public",
-    ) {
-
-    override fun createContainer(): PostgreSQLContainer<*> {
-        val dockerImageName: DockerImageName =
-            DockerImageName.parse(DOCKER_IMAGE_NAME).asCompatibleSubstituteFor(DOCKER_IMAGE_NAME)
-        val modifier: TestContainerFactory.ContainerModifier<PostgreSQLContainer<*>> =
-            TestContainerFactory.newModifier("withWalLevelLogical") {
-                it.withCommand("postgres -c wal_level=logical")
-            }
-        return TestContainerFactory.exclusive(dockerImageName, modifier)
-    }
-
-    companion object {
-        const val DOCKER_IMAGE_NAME = "postgres:13-alpine"
-        init {
-            TestContainerFactory.register(DOCKER_IMAGE_NAME, ::PostgreSQLContainer)
-        }
-        const val PUBLICATION_NAME = "test_publication"
-        const val SLOT_NAME = "test_slot"
-    }
-
-    override fun PostgreSQLContainer<*>.createStream() {
-        withStatement { it.execute("CREATE TABLE tbl (id SERIAL PRIMARY KEY, v INT)") }
-        withStatement { it.execute("ALTER TABLE tbl REPLICA IDENTITY FULL") }
-        withStatement {
-            it.execute("SELECT pg_create_logical_replication_slot('$SLOT_NAME', 'pgoutput')")
-        }
-        withStatement { it.execute("CREATE PUBLICATION $PUBLICATION_NAME FOR TABLE tbl") }
-    }
-
-    override fun PostgreSQLContainer<*>.insert12345() {
-        for (i in 1..5) {
-            withStatement { it.execute("INSERT INTO tbl (v) VALUES ($i)") }
-        }
-    }
-
-    override fun PostgreSQLContainer<*>.update135() {
-        withStatement { it.execute("UPDATE tbl SET v = 6 WHERE id = 1") }
-        withStatement { it.execute("UPDATE tbl SET v = 7 WHERE id = 3") }
-        withStatement { it.execute("UPDATE tbl SET v = 8 WHERE id = 5") }
-    }
-
-    override fun PostgreSQLContainer<*>.delete24() {
-        withStatement { it.execute("DELETE FROM tbl WHERE id = 2") }
-        withStatement { it.execute("DELETE FROM tbl WHERE id = 4") }
-    }
-
-    private fun <X> PostgreSQLContainer<*>.withStatement(fn: (Statement) -> X): X =
-        createConnection("").use { connection: Connection ->
-            connection.createStatement().use { fn(it) }
-        }
-
-    override fun position(recordValue: DebeziumRecordValue): LogSequenceNumber? {
-        val lsn: Long =
-            recordValue.source["lsn"]?.takeIf { it.isIntegralNumber }?.asLong() ?: return null
-        return LogSequenceNumber.valueOf(lsn)
-    }
-
-    override fun position(sourceRecord: SourceRecord): LogSequenceNumber? {
-        val offset: Map<String, *> = sourceRecord.sourceOffset()
-        val lsn: Long = offset["lsn"] as? Long ?: return null
-        return LogSequenceNumber.valueOf(lsn)
-    }
-
-    override fun PostgreSQLContainer<*>.currentPosition(): LogSequenceNumber =
-        withStatement { statement: Statement ->
-            statement.executeQuery("SELECT pg_current_wal_lsn()").use {
-                it.next()
-                LogSequenceNumber.valueOf(it.getString(1))
-            }
-        }
-
-    override fun PostgreSQLContainer<*>.syntheticInput(): DebeziumInput {
-        val (position: LogSequenceNumber, txID: Long) =
-            withStatement { statement: Statement ->
-                statement.executeQuery("SELECT pg_current_wal_lsn(), txid_current()").use {
-                    it.next()
-                    LogSequenceNumber.valueOf(it.getString(1)) to it.getLong(2)
-                }
-            }
-        val timestamp: Instant = Instant.now()
-        val key: ArrayNode =
-            Jsons.arrayNode().apply {
-                add(databaseName)
-                add(Jsons.objectNode().apply { put("server", databaseName) })
-            }
-        val value: ObjectNode =
-            Jsons.objectNode().apply {
-                put("ts_usec", timestamp.toEpochMilli() * 1000L)
-                put("lsn", position.asLong())
-                put("txId", txID)
-            }
-        val offset = DebeziumOffset(mapOf(key to value))
-        val state = DebeziumState(offset, schemaHistory = null)
-        val syntheticProperties: Map<String, String> = debeziumProperties()
-        return DebeziumInput(syntheticProperties, state, isSynthetic = true)
-    }
-
-    override fun PostgreSQLContainer<*>.debeziumProperties(): Map<String, String> =
-        DebeziumPropertiesBuilder()
-            .withDefault()
-            .withConnector(PostgresConnector::class.java)
-            .withDebeziumName(databaseName)
-            .withHeartbeats(heartbeat)
-            .with("plugin.name", "pgoutput")
-            .with("slot.name", SLOT_NAME)
-            .with("publication.name", PUBLICATION_NAME)
-            .with("publication.autocreate.mode", "disabled")
-            .with("flush.lsn.source", "false")
-            .withDatabase("hostname", host)
-            .withDatabase("port", firstMappedPort.toString())
-            .withDatabase("user", username)
-            .withDatabase("password", password)
-            .withDatabase("dbname", databaseName)
-            .withOffset()
-            .withStreams(listOf(stream))
-            .buildMap()
-}
-
-class CdcPartitionReaderMongoTest :
-    CdcPartitionReaderTest<BsonTimestamp, MongoDbReplicaSet>(
-        namespace = "test",
-    ) {
-
-    override fun createContainer(): MongoDbReplicaSet {
-        return MongoDbReplicaSet.replicaSet().memberCount(1).build().also {
-            for (hostName in it.hostNames) {
-                DnsCacheManipulator.setDnsCache(hostName, "127.0.0.1")
-            }
-            it.start()
-        }
-    }
-
-    override fun MongoDbReplicaSet.createStream() {
-        withMongoCollection { it.insertOne(Document("_id", 0)) }
-    }
-
-    override fun MongoDbReplicaSet.insert12345() {
-        withMongoCollection {
-            for (i in 1..5) {
-                it.insertOne(Document("_id", i).append("v", i))
-            }
-        }
-    }
-
-    override fun MongoDbReplicaSet.update135() {
-        withMongoCollection {
-            it.updateOne(Document("_id", 1), Updates.set("v", 6))
-            it.updateOne(Document("_id", 3), Updates.set("v", 7))
-            it.updateOne(Document("_id", 5), Updates.set("v", 8))
-        }
-    }
-
-    override fun MongoDbReplicaSet.delete24() {
-        withMongoCollection {
-            it.deleteOne(Document("_id", 2))
-            it.deleteOne(Document("_id", 4))
-        }
-    }
-
-    private fun <X> MongoDbReplicaSet.withMongoClient(fn: (MongoClient) -> X): X =
-        MongoClients.create(connectionString).use { fn(it) }
-
-    private fun <X> MongoDbReplicaSet.withMongoDatabase(fn: (MongoDatabase) -> X): X =
-        withMongoClient {
-            fn(it.getDatabase(stream.namespace!!))
-        }
-
-    fun <X> MongoDbReplicaSet.withMongoCollection(fn: (MongoCollection<Document>) -> X): X =
-        withMongoDatabase {
-            fn(it.getCollection(stream.name))
-        }
-
-    override fun position(recordValue: DebeziumRecordValue): BsonTimestamp? {
-        val resumeToken: String =
-            recordValue.source["resume_token"]?.takeIf { it.isTextual }?.asText() ?: return null
-        return ResumeTokens.getTimestamp(ResumeTokens.fromData(resumeToken))
-    }
-
-    override fun position(sourceRecord: SourceRecord): BsonTimestamp? {
-        val offset: Map<String, *> = sourceRecord.sourceOffset()
-        val resumeTokenBase64: String = offset["resume_token"] as? String ?: return null
-        return ResumeTokens.getTimestamp(ResumeTokens.fromBase64(resumeTokenBase64))
-    }
-
-    override fun MongoDbReplicaSet.currentPosition(): BsonTimestamp =
-        ResumeTokens.getTimestamp(currentResumeToken())
-
-    override fun MongoDbReplicaSet.syntheticInput(): DebeziumInput {
-        val resumeToken: BsonDocument = currentResumeToken()
-        val timestamp: BsonTimestamp = ResumeTokens.getTimestamp(resumeToken)
-        val resumeTokenString: String = ResumeTokens.getData(resumeToken).asString().value
-        val key: ArrayNode =
-            Jsons.arrayNode().apply {
-                add(stream.namespace)
-                add(Jsons.objectNode().apply { put("server_id", stream.namespace) })
-            }
-        val value: ObjectNode =
-            Jsons.objectNode().apply {
-                put("ord", timestamp.inc)
-                put("sec", timestamp.time)
-                put("resume_token", resumeTokenString)
-            }
-        val offset = DebeziumOffset(mapOf(key to value))
-        val state = DebeziumState(offset, schemaHistory = null)
-        val syntheticProperties: Map<String, String> = debeziumProperties()
-        return DebeziumInput(syntheticProperties, state, isSynthetic = true)
-    }
-
-    private fun MongoDbReplicaSet.currentResumeToken(): BsonDocument =
-        withMongoDatabase { mongoDatabase: MongoDatabase ->
-            val pipeline = listOf<Bson>(Aggregates.match(Filters.`in`("ns.coll", stream.name)))
-            mongoDatabase.watch(pipeline, BsonDocument::class.java).cursor().use {
-                it.tryNext()
-                it.resumeToken!!
-            }
-        }
-
-    override fun MongoDbReplicaSet.debeziumProperties(): Map<String, String> =
-        DebeziumPropertiesBuilder()
-            .withDefault()
-            .withConnector(MongoDbConnector::class.java)
-            .withDebeziumName(stream.namespace!!)
-            .withHeartbeats(heartbeat)
-            .with("capture.scope", "database")
-            .with("capture.target", stream.namespace!!)
-            .with("mongodb.connection.string", connectionString)
-            .with("snapshot.mode", "no_data")
-            .with(
-                "collection.include.list",
-                DebeziumPropertiesBuilder.joinIncludeList(
-                    listOf(Pattern.quote("${stream.namespace!!}.${stream.name}"))
-                )
-            )
-            .with("database.include.list", stream.namespace!!)
-            .withOffset()
-            .buildMap()
-
-    override fun deserialize(
-        key: DebeziumRecordKey,
-        value: DebeziumRecordValue,
-        stream: Stream,
-    ): DeserializedRecord {
-        val id: Int = key.element("id").asInt()
-        val record: Record =
-            if (value.operation == "d") {
-                Delete(id)
-            } else {
-                val v: Int? =
-                    value.after
-                        .takeIf { it.isTextual }
-                        ?.asText()
-                        ?.let { Jsons.readTree(it)["v"] }
-                        ?.asInt()
-                if (v == null) {
-                    // In case a mongodb document was updated and then deleted, the update change
-                    // event will not have any information ({after: null})
-                    // We are going to treat it as a Delete.
-                    Delete(id)
-                } else if (value.operation == "u") {
-                    Update(id, v)
-                } else {
-                    Insert(id, v)
-                }
-            }
-        return DeserializedRecord(
-            data = Jsons.valueToTree(record),
-            changes = emptyMap(),
-        )
     }
 }
